backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyYAML==6.0
python-dotenv==1.0.0
networkx==3.2
pydantic==2.5.0
Jinja2==3.1.2
